Avoid warnings with shortening of 64bit to 32bit
Technically this code is limiting the containers to 2B objects or less,
a more appropriate fix might be to return an appropriately signed
version of size_t

+++ json.hpp
@@ -265,7 +265,7 @@ class JSON
 
         int length() const {
             if( Type == Class::Array )
-                return Internal.List->size();
+                return static_cast<int>(Internal.List->size());
             else
                 return -1;
         }
@@ -278,9 +278,9 @@ class JSON
 
         int size() const {
             if( Type == Class::Object )
-                return Internal.Map->size();
+                return static_cast<int>(Internal.Map->size());
             else if( Type == Class::Array )
-                return Internal.List->size();
+                return static_cast<int>(Internal.List->size());
             else
                 return -1;
         }
